Refactor client dashboard navigation
