+++ app/client/components/ClientShell.module.css
@@ -6,17 +6,6 @@
     var(--color-surface-subtle);
 }
 
-.body {
-  display: flex;
-  flex: 1;
-  width: 100%;
-  gap: var(--space-2);
-}
-
-.sidebar {
-  display: none;
-}
-
 .main {
   flex: 1;
   padding: var(--space-6) clamp(var(--space-4), 6vw, var(--space-8));
@@ -26,16 +15,6 @@
 }
 
 @media (min-width: 960px) {
-  .body {
-    padding-inline: var(--space-6);
-  }
-
-  .sidebar {
-    display: block;
-    width: 220px;
-    flex-shrink: 0;
-  }
-
   .main {
     margin-left: 0;
     max-width: none;
@@ -45,15 +24,6 @@
 }
 
 @media (min-width: 1200px) {
-  .body {
-    gap: var(--space-6);
-    padding-inline: var(--space-8);
-  }
-
-  .sidebar {
-    width: 240px;
-  }
-
   .main {
     padding-inline: var(--space-8);
   }
