--- app/client/components/DashboardSidebar.module.css
@@ -1,82 +1,0 @@
-.sidebar {
-  position: sticky;
-  top: 76px;
-  display: none;
-  flex-direction: column;
-  gap: var(--space-2);
-}
-
-.sectionLabel {
-  font-size: 0.75rem;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.08em;
-  color: var(--color-text-soft);
-  margin-bottom: var(--space-1);
-}
-
-.navList {
-  display: grid;
-  gap: var(--space-1);
-}
-
-.navItem {
-  display: flex;
-  align-items: center;
-  gap: var(--space-2);
-  padding: var(--space-2) var(--space-3);
-  border-radius: var(--radius-md);
-  color: var(--color-text-soft);
-  transition: background 0.2s ease, color 0.2s ease, transform 0.2s ease;
-}
-
-.navItem:hover {
-  color: var(--color-text);
-  background: rgba(148, 163, 184, 0.14);
-}
-
-.navItemActive {
-  color: var(--color-text);
-  background: rgba(37, 99, 235, 0.14);
-  box-shadow: inset 0 0 0 1px rgba(37, 99, 235, 0.32);
-}
-
-.navItemIcon {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  border-radius: 12px;
-  background: var(--color-surface);
-  box-shadow: var(--shadow-ring);
-  color: var(--color-text-muted);
-  flex-shrink: 0;
-}
-
-.navItemLabel {
-  font-size: 0.9rem;
-  font-weight: 500;
-}
-
-@media (min-width: 960px) {
-  .sidebar {
-    display: flex;
-  }
-}
-
-@media (max-width: 1120px) {
-  .navItemLabel {
-    display: none;
-  }
-
-  .sidebar {
-    align-items: center;
-  }
-}
-
-@media (min-width: 1121px) {
-  .sidebar {
-    align-items: stretch;
-  }
-}
